// src/api/generated/models/listInternalAppointmentsApiV1AppointmentsInternalGetParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Community App V2
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentSortField } from './appointmentSortField'
import type { AppointmentStatus } from './appointmentStatus'
import type { SortOrder } from './sortOrder'

export type ListInternalAppointmentsApiV1AppointmentsInternalGetParams = {
  office_id?: string | null
  citizen_id?: string | null
  ticket_id?: string | null
  status?: AppointmentStatus | null
  sort_by?: AppointmentSortField
  order?: SortOrder
  starts_from?: string | null
  starts_to?: string | null
  created_from?: string | null
  created_to?: string | null
  /**
   * @minimum 1
   */
  page?: number
  /**
   * @minimum 1
   * @maximum 100
   */
  size?: number
  q?: string | null
}